+++ include/Lektra.hpp
@@ -207,7 +207,7 @@ private:
                           std::function<void()> onAllDone) noexcept;
     void focusSplitHelper(DocumentContainer::Direction direction) noexcept;
     void splitHelper(Qt::Orientation orientation) noexcept;
-    bool checkConfigFile(const QString &path) const noexcept;
+    bool checkConfigFile(const QString &path) noexcept;
 
     DocumentView *
     openFileSplitHelper(const QString &filename               = {},